Создал дополнительный модуль create-photo-album
